requirements.txt
astroid==2.2.3
colorama==0.4.1
Django==2.1.7
djangorestframework==3.9.2
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
